[FIX] point_of_sale, pos_restaurant: fix random error in test
The `PreparationPrinterContent` test was checking if the preparation
receipt was containing a specific string. But the check was performed
on the `.render-container` which is generated for printing and erased
just after. That was causing a random error in the test.

Now we call internal `pos_store` method that generate the receipt
content and we check if the string is present in the content.

runbot error id: 114884

closes odoo/odoo#197681

X-original-commit: e83cb544e6424cffba29811b33f705a5bd2cad0c
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: David Monnom (moda) <moda@odoo.com>